tls/client auth: verify all certificates in client request
When client certificate is enabled Caddy check only last certificate from
request. When this cert is not in list of trusted leaf certificates,
connection is rejected. But in some cases client certificate is not
last certificate in list; sometimes is first, before certificates of CA.

This patch add checking all sent certificates - if any match with trusted
certificates connection is accepted.